requirements.txt
django==1.11
django-appconf==1.0.1
django-image-cropping==1.1.0
easy-thumbnails==2.5
mysqlclient==1.3.12
pillow==5.0.0
pytz==2018.3
six==1.11.0
gunicorn==19.7.1